+++ backend/src/main/java/com/internproject/premium_cafe_backend/service/serviceimpl/CategoryServiceImpl.java
@@ -3,9 +3,12 @@ package com.internproject.premium_cafe_backend.service.serviceimpl;
 import com.internproject.premium_cafe_backend.dto.request.CategoryRequestDto;
 import com.internproject.premium_cafe_backend.dto.response.CategoryResponseDto;
 import com.internproject.premium_cafe_backend.entity.Category;
+import com.internproject.premium_cafe_backend.exception.DuplicateResourceException;
+import com.internproject.premium_cafe_backend.exception.InvalidRequestException;
 import com.internproject.premium_cafe_backend.exception.ResourceNotFoundException;
 import com.internproject.premium_cafe_backend.mapper.CategoryMapper;
 import com.internproject.premium_cafe_backend.repository.CategoryRepository;
+import com.internproject.premium_cafe_backend.repository.MenuItemRepository;
 import com.internproject.premium_cafe_backend.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,12 +21,14 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final MenuItemRepository menuItemRepository;
 
     @Override
     public CategoryResponseDto createCategory(CategoryRequestDto request) {
 
         if (categoryRepository.findByName(request.getName()).isPresent()) {
-            throw new RuntimeException("Category already exists.");
+            throw new DuplicateResourceException(
+                    "Category already exists with name : " + request.getName());
         }
         Category category = CategoryMapper.toEntity(request);
 
@@ -58,6 +63,14 @@ public class CategoryServiceImpl implements CategoryService {
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Category not found with id : " + id));
 
+        if (!category.getName().equals(request.getName())
+                && categoryRepository.findByName(request.getName()).isPresent()) {
+
+            throw new DuplicateResourceException(
+                    "Category already exists with name : "
+                            + request.getName());
+        }
+
         category.setName(request.getName());
         category.setDescription(request.getDescription());
         category.setUpdatedAt(LocalDateTime.now());
@@ -74,6 +87,10 @@ public class CategoryServiceImpl implements CategoryService {
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Category not found with id : " + id));
 
+        if (menuItemRepository.existsByCategory(category)) {
+            throw new InvalidRequestException(
+                    "Cannot delete category because menu items are associated with it.");
+        }
         categoryRepository.delete(category);
     }
 }
